backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
psutil==5.9.6
python-jose==3.3.0
passlib==1.7.4
pytz==2023.3
scipy==1.11.4
pyspark==3.5.0
aiofiles==23.2.1
